requirements.txt
google-generativeai>=0.5.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
pytrends>=4.9.2
Pillow>=10.0.0
requests>=2.31.0
